Validate and select deployment environment at run time
This commit adds steps to validate and select the target deployment
environment and map it to the appropriate domain name. Contrary to
earlier testing, it seems like the default of 'preview' doesn't
propagate.

Validation must be one of: {dev,staging,prod,preview}

Replace ${{ inputs.environment }} with $DEPLOYMENT_ENV

The document was also formatted for tab spacing and line length.